Unix file descriptors: add support for custom mmap implementations
This commit adds an mmap() closure to the file descriptor struct,
which allows an open file descriptor to have a custom mmap
implementation. The io_uring code has been refactored to use the
mmap closure, and as a result the io_uring-specific code in mmap.c
could be replaced by more generic code.
The MAP_FIXED mmap flag is now supported for custom mappings; in
order to prevent modifications to existing mappings when a given
mmap() syscall with the MAP_FIXED flag fails, the unmapping of
address ranges that overlap with the fixed range is not done for
custom mappings; instead, a custom mmap implementation is expected
to overwrite any existing overlapping mappings; for this purpose,
the remap() function has been added. The map_level() function now
takes an additional parameter (a flush entry) which indicates
whether an existing mapping should be overwritten (in which case
the overwritten pages are added to the flush entry to be
subsequently invalidated).
When a non-fixed mapping is requested on a file descriptor with a
custom mmap, first a suitable virtual address range is identified
via process_get_virt_range_locked(), then the mmap closure is
invoked, and only if the closure function returns successfully the
previously identified address range is added to the process vmaps.
Custom mappings are not treated as pre-allocated virtual address
ranges; instead, they are unmapped like any other mapping type.
This means that the io_uring code should not tear down any existing
mapping for a given file descriptor when the file descriptor is
closed (it has been verified that on Linux an mmapped memory range
is still accessible after closing an io_uring file descriptor);
instead, mappings on a file descriptor should increment the
reference count of the file descriptor, which prevents its
associated resources from being deallocated. To implement this, the
fsfile pointer in the vmap structure has been replaced by a more
generic file descriptor pointer, so that reference counting is done
at the file descriptor level.