Runtime tables: add table_remove() function
This function allows to retrieve and remove an entry from a table
with a single lookup. It can be used in place of a
`table_find(); table_set();` call sequence, which would look up
twice the element to be removed.
The file_unlink() function in the TFS code has been changed to use
table_remove().