filesystem_resolve_cstring(): return FS_STATUS_NOENT on empty path
This change is needed in order for syscalls such as unlink() to
return -ENOENT when an empty path argument is supplied.
The unlink runtime test has been amended to test this behavior.

This change has a side effect that makes a glibc bug
(https://bugs.launchpad.net/ubuntu/+source/glibc/+bug/1982326) more
invasive when running the thread runtime test program (even though
filesystem_resolve_cstring() is never called with an empty path
argument when running this program), which triggers a test failure
in Jenkins. As a workaround, the sleep time in the thread test
program has been increased.