Handle empty queue when dequeuing signal
If dequeue_signal is working on the process signal queue, then multiple threads
may be trying to access that queue simultaneously in an smp machine. The
pending check was happening outside the queue lock which could result in
a thread trying to dequeue an already processed signal from a now empty
queue, resulting in an assertion failure.

This changes the assert in sigstate_dequeue_signal() to be a check that
returns INVALID_ADDRESS if the queue is empty. A new smp signal test for
this condition has been added to the signal runtime tests.