Unix process: fix calculation of CPU times
If the CPU time of a process is calculated from the list of active
threads, the CPU time from terminated threads is not accounted for.
This change fixes the issue by adding two timestamps to the process
struct, where the (system and user) CPU time accumulated from all
(past and present) threads is stored.