Unix timers: add support for process and thread CPU time
This changeset adds support for CLOCK_PROCESS_CPUTIME_ID and
CLOCK_THREAD_CPUTIME_ID to the Unix timers created via the
timer_create syscall; in addition, ITIMER_PROF is now supported by
the getitimer and setitimer syscalls.
The timerqueue code has been enhanced so that it can be used with
clock types (such as process and thread CPU time) that are not
derived from the monotonic system clock. The Unix process struct
now has its own timerqueue, which handles timers that track the
total CPU time spent by a process. Analogously, the thread struct
has a timerqueue that handles timers tracking CPU time spent by a
single thread (the thread timerqueue is only instantiated if/when a
timer that requires it is created).
The Unix timer code uses the appropriate timerqueue based on the
type of clock used by a given timer.
Process and thread timerqueues are serviced when the CPU time spent
by a thread is updated; this means that a timer based on a thread
CPU time may fire with a delay up to RUNLOOP_TIMER_MAX_PERIOD_US
microseconds (assuming execution of a syscall does not take longer
than that), and a timer based on the process CPU time may fire with
a delay up to RUNLOOP_TIMER_MAX_PERIOD_US * NCPU) microseconds,
where NCPU is the number of vCPUs of the VM.
The timer runtime test has been amended to exercise the newly
supported timers.

Closes #1787