syscall_accumulate_stime(): do not access terminated threads
When the exit() syscall is invoked, the calling thread is
terminated and its structure is deallocated. Therefore, the thread
pointer in the syscall_context structure points to an invalid
thread structure. When the syscall context is paused, if the
syscall execution time is added to the thread system time, an
invalid memory access occurs.
This change fixes the above issue by zeroing the thread pointer in
the syscall context when a thread is terminated, and checking for a
non-null pointer before adding the syscall execution time to the
thread system time.