KVM: fix detection logic when Hyper-V Enlightenments are enabled
When running under QEMU with KVM acceleration, the hypervisor
signature reported at CPUID function 0x40000000 may not be the KVM
signature; specifically, if Hyper-V Enlightenments are enabled in
the QEMU command line (for example, with the command line option
`--cpu host,hv_relaxed`), the hypervisor reported at CPUID function
0x40000000 is Hyper-V, whereas KVM is reported at CPUID function
0x40000100.
This change fixes the hypervisor detection logic so that KVM is
properly detected when Hyper-V Enlightenments are enabled. In
addition, the check on the eax value (which indicates the maximum
cpuid function present in a given leaf) is being fixed so that the
detection works properly on VMs (such as AWS c5 instances) where
additional cpuid functions (beyond KVM_CPUID_FEATURES) are present.

Closes #1791.