ELF dynamic linking: add support for multiple relocation sections
In an ELF file there may be more than one relocation section, for
example to differentiate between lazily non-lazily bound symbols.
When there are multiple relocation sections, they refer to
(portions of) a single contiguous array of relocation entries.
This change refactors the dynamic linking code by adding a generic
function elf_dyn_parse() that parses the dynamic section of an ELF
file and returns the tables of interest; the relocation table is
not returned as a section header (which may not cover all the
relocations), but as an array of relocation entries.
In addition, the sh_addr field from section headers is used instead
of sh_offset when matching a section with its reference from a
dynamic tag; this fixes parsing of files where file offsets do not
match virtual addresses.