Add ltrace functionality
This feature allows tracing calls made by the application binary to
dynamic library functions. It is enabled by adding an 'ltrace'
attribute to the root tuple, and prints on the serial console
(similarly to the Linux ltrace tool) a line for each dynamic
library function call (with the function name) and for each signal
delivered to the process (with the signal number).
This feature is implemented for the x86 and aarch64 architectures
(in riscv, CPU debug registers cannot be accessed from supervisor
mode, and the SBI does not currently have the functionality to
access debug registers via SBI calls).
The elf_dyn_parse() and elf_plt_get() functions are used by ltrace
to retrieve the information needed to be able to place software
breakpoints where dynamic function calls are made and map each
breakpoint to the corresponding function.
The Unix fault handler has been changed so that trap exceptions
generated after a single-stepped instruction are correctly handled
(either by ltrace, or by generating a SIGTRAP signal).