virtio-net TX completion closures: use dedicated objcache
Analyzing lock stats while stress-testing a Go web server running
on a multi-core GCP instance revealed a high contention of the lock
that protects the kernel locked heap during allocation and
deallocation of TX completion closures in the virtio-net driver.
Using a dedicated objcache for these closures instead of the locked
heap decreases the lock contention and improves the performance of
the web server.