Unix: epoll: replace fds lock with read-write lock
Analyzing lock stats while stress-testing a Go web server running
on a multi-core instance revealed a high contention of the lock
that protects the file descriptor list associated to an epoll
instance. Since this list is only modified when adding or removing
a file descriptor to/from an epoll instance (and not when invoking
the epoll_wait syscall, which happens more frequently), replacing
the lock with a read-write lock decreases the lock contention and
improves the performance of the web server when servicing multiple
connections simultaneously.