Klibs: AWS CloudWatch: refactor code for sending memory metrics
This refactoring is being done in order to avoid code duplication
in the next commit, which will add support for CloudWatch logs.
The new utility functions aws_req_set_date() and cw_aws_req_send()
are being defined and used to send memory metrics; the code for DNS
resolution is being changed in order to use generic functions (such
as cw_connect, cw_dns_cb and cw_send_async) which will be reused in
the next commit.