Klibs: AWS CloudWatch: add check for AWS metadata availability
Sometimes it may take a few seconds for an AWS instance to receive
an IP address from the DHCP server. During this interval, the AWS
metadata server at address 169.254.169.254 is not reachable, which
prevents the cloudwatch klib from retrieving the metadata necessary
for its operation. When this happens, the current code prints on
the console, as a consequence of being unable to connect to the
metadata server, a series of error messages, among which there is a
misleading "no IAM role associated to instance" message.
To avoid printing misleading error messages, this commit adds a
check for reachability of the metadata server before trying to
connect to the server; if this check fails, a more appropriae "AWS
metadata not available" error message is printed, and the
initialization is retried after a backoff delay.