Flush waiters upon shutdown of udp socket
If a thread is blocked on a recv on a udp socket, then calling shutdown on
that socket from another thread should wake up the blocked thread to abort
the recv. The return value from the recv should be 0 to indicate an end-of-file
condition and not ERESTARTSYS, so an additional shutdown state is added
to udp sockets and set on shutdown. This is used by the bh handler to
decide whether to return 0 or ERESTARTSYS, matching the existing tcp
socket behavior.

A runtime test has also been added to test this condition and return values.