open(2): add support for O_TRUNC flag
When this flag is specified in an open() or openat() syscall, if
the file being opened already exists and is a regular file it is
truncated to length 0.
To implement this feature, the filesystem_get_node() function has
been amended to take an additional 'truncate' boolean argument;
the fsfile_open_or_create() function (which calls
filesystem_get_node) has also been amended to take an additional
'truncate' argument, and this allowed to remove calls to
fsfile_truncate() in a few places.

Closes #1803.