netsock_shutdown(): fix handling of multiple calls on a socket
The shutdown syscall can be invoked on a given socket multiple
times; if it is invoked once with `how` set to SHUT_RD and once
with `how` set to SHUT_RW, at the second call both directions of
the underlying TCP connection are shut down, which means that the
lwIP tcp_pcb structure can be deallocated at any time and is
therefore unsafe to reference. The existing code is failing to
detect this condition, which may cause lwIP functions to be called
on a tcp_pcb structure that has already been deallocated.
This change fixes the above issue. In addition, the initial check
on the state member of the socket structure is now performed with
the lwIP lock held, for multi-thread safety.