volume_add(): take generic fs_init_handler parameter
Instead of having the volume_add() function take a
storage_req_handler which is then used to create a TFS filesystem
when the volume is mounted, the volume_add() function now takes a
more generic fs_init_handler parameter, which is a closure that is
then applied when the volume is mounted. This will allow handling
non-TFS volumes.
A failure to initialize a filesystem is properly handled even if
no valid filesystem pointer is passed to the fs_init_handler
completion.
A generic private pointer is now used as volume identifier (which
is needed for volume removal) instead of the storage_req_handler
parameter.