Page cache: touch_or_fill_page_nodelocked(): fix page refcount
If the touch_or_fill_page_nodelocked() function is called with a
null merge argument, the caller does not want to trigger a read
request to fill the page, and does not want to be notified when the
page is filled; in these conditions, if a page is not filled, the
page refcount should not be incremented, because the caller does
not have a code path where the refcount will be decremented.
This change fixes a stray refcount increment that happens when
pagecache_map_page_if_filled() is called for a page in one of the
{FREE, ALLOC, READING} states, which, if the relevant page cache
node is later deallocated, causes an assertion failure in
pagecache_page_release() where the page state will not be FREE as
expected.