TFS fsfile: purge pagecache node when deallocating fsfile
When a TFS fsfile is deallocated, it means the underlying file is
no longer part of the filesystem; for this reason, it is not
necessary to sync to the storage medium any dirty pages in the page
cache for this file.
This change adds a new pagecache_purge_node() function in the page
cache code, and modifies the fsfile refcount completion closure to
call pagecache_purge_node() instead of pagecache_sync_node().
Beside avoiding unnecessary writes to disk, this change will make
it possible to decrement the fsfile reference count while holding
the filesytem lock (because storage write requests are no longer
triggered by the refcount completion closure).