TFS: use a separate lock for the storage heap
This change adds a new spinlock to the TFS code to protect the
storage heap from concurrent access; this allows using the storage
heap without taking the filesystem lock. As a result, the
filesystem lock does no longer need to be held when an fsfile is
deallocated.
This change is done in preparation for the addition of a generic
filesystem abstraction layer.