Filesystem: add support for 9P, using virtIO-9P devices
This change adds support for the 9P filesystem exported by
virtIO-9P devices, which allows the kernel to access files on the
host filesystem (which can be made accessible to the guest by QEMU
with the `-virtfs` command line option).
The numerical value of the `mount_tag` flag in the virtfs command
line option is reflected in the attachment point that is associated
to the volume created by the kernel when the relevant virtIO-9P
device is probed, so that it is possible to instruct the kernel to
mount a 9P filesystem by providing a
`mounts:(%<attachment_id>: <mount_dir>)` directive in the manifest.