Add option to run user program after getting ip4 address
Some user programs assume the network is ready to go when executing, but
if dhcp takes a long time then the user program may hit network errors
if an address and gateway isn't acquired yet. This change adds a new manifest
option 'exec_wait_for_ip4_secs', which will delay running the user program
until either an ip4 address is acquired or the timeout specified in the option
has elapsed. If a static address is set then the program is executed immediately.