CircleCI: nightly-build-mac: remove `brew upgrade python` command
In the CircleCI Mac container, Python is not installed by default,
thus the `brew upgrade python` command fails with "Error: python
not installed".
This change removes the above command, which is not needed because
the Google Cloud client installer (which is run in the following
steps) automatically installs Python 3.7 (and would use this Python
installation even if another Python version was already installed).