Unit tests: parser: refactor code to allow adding non-tuple parsers
This change refactors the tuple parser unit test code so that other
(non-tuple) parsers can be added to the test suite; the next commit
will add test cases for the JSON parser.
In addition, the string passed to the parser error closure is now
properly copied (and deallocated when no longer used).