Runtime: add JSON parser implementation
This change implements a JSON parser and adds a set of test cases
to exercise the parser in the unit tests.
The parser converts a JSON string to a tuple, where each
string-valued JSON attribute correspondes to a tuple element with a
string value, and each object-valued attribute corresponds to a
tuple element with a tuple value. Other JSON attribute types
(number, boolean, array and null) are currently not supported (the
parser is able to parse a JSON string containing these attribute
types, but does not convert them to tuple elements).
In order to properly deallocate the memory allocated when parsing
an incomplete JSON string, the CHARACTER_INVALID special character
is passed to the parser to signal the end of an input string: if
the input string is invalid, the end of the string triggers a call
to the parser error closure.