Klibs: cloud_init: refactor file download code
This refactoring is made in preparation for adding support for
cloud_init tasks different from file downloads.
A generic cloud_init_task closure is being defined; this closure is
invoked with an opcode (currently defined opcodes are START and
DELETE) and a generic pointer argument. When parsing the cloud_init
configuration specified in the manifest, tasks are allocated and
added to a vector; then, if all configuration items are valid, all
tasks from the vector are started, otherwise all the tasks are
deallocated and the klib returns an error.
The cloud_download_cfg structure is being stripped of file-specific
fields, which are now in the cloud_download_file structure. The
file download code is being reorganized to separate it from generic
code that can be reused for other cloud_init tasks.