Klibs: cloud_init: add support for environment variables
This change adds to the cloud_init klib support for populating the
environmemt of the user process with variables whose name and value
are retrieved from an HTTP(S) server during startup.
The configuration syntax in the manifest to use this functionality
is as in the following example:
```
download_env:[
  (
    src: "http://10.0.2.2:8200/v1/secret/data/hello"
    auth: "Bearer hvs.6v6yY5yZf32uZzaa7HhiX6AZ"
    path: "attr1/attr2"
  )
]
```
The `download_env` attribute is an array where each element
specifies a download source URL, an optional authentication header,
and an optional attribute path.
For each element in the `download_env` attribute, the klib executes
an HTTP request to the specified URL, and if the peer responds
successfully, the response body is parsed as a JSON object, and
from this object an "environment object" is retrieved. If the
attribute path (i.e. the `path` element in the manifest) is not
present, the environment object corresponds to the root JSON object
of the response body; otherwise, for each element in the attribute
path (where the element separator is the / character), a nested
JSON object is retrieved from the response body by looking up an
attribute named after the element (the corresponding attribute
value must be a JSON object): the environment object is the nested
object corresponding to the last element of the attribute path.
Once the environment object is identified, all string-valued
attributes in this object are converted to environment variables
(any non-string-valued attributes are ignored).
Examples:
- an empty attribute path can be used to retrieve the environment
variables from the following response body:
```
{
  "VAR1": "value1",
  "VAR2": "value2"
}
```
- an attribute path set to "obj1/obj2" can be used to retrieve the
environment variables from the following response body:
```
{
  "obj1: {
    "obj2": {
      "VAR1": "value1",
      "VAR2": "value2"
    }
  }
}
```

Closes #1813.