Sockets: move standard definitions to header files
Definitions of constants such as SIOCG* and structures such as
struct mmsghdr are not an implementation detail of the kernel
socket code, thus they can be moved to header files so that other
code has access to them.
This change (which also inserts additional socket-related constant
definitions) is being done in preparation for adding the sandbox
klib.