Klibs: add sandbox klib
This klib implements a sandboxing mechanism by which the
capabilities of the running process can be restricted by overriding
certain syscall handlers. The current implementation supports the
OpenBSD pledge (https://man.openbsd.org/pledge.2) and unveil
(https://man.openbsd.org/unveil.2) syscalls.

A pledge invocation forces the process into a restricted-service
operating mode by limiting the set of syscalls (and in some cases
the allowed values of syscall arguments) that can be invoked.
Syscalls that are present in both Nanos and OpenBSD are handled by
following the pledge implementation in OpenBSD, except for syscalls
that are not applicable to unikernels (e.g. those related to inter-
process communication). Syscalls that are present (or could be
implemented in the future) in Nanos but not in OpenBSD are handled
in the same way as "similar" or related OpenBSD syscalls when such
syscalls could be found, and are always forbidden when no related
syscalls could be found in OpenBSD.
The pledge syscall can be invoked in a C program by calling the
`pledge()` function from the following code:
```
int pledge(const char *promises, const char *execpromises)
{
    return syscall(335, promises, execpromises);
}
```

The unveil syscall allows adding and removing (or restricting)
visibility of filesystem paths for the running process. It can be
invoked in a C program by calling the `unveil()` function from
the following code:
```
int unveil((const char *path, const char *permissions)
{
    return syscall(336, path, permissions);
}
```

The pledge and unveil features are enabled in the sandbox klib if
the manifest contains a "sandbox" tuple with a "pledge" and
"unveil" tuple, respectively.
A new runtime test program is being added to exercise these
features.