netlink recvfrom/recvmsg: don't set source address in error cases
If a recvfrom() or recvmsg() call on a netlink socket results in an
error (e.g. when no messages have been received and the socket is
non-blocking), the source address and corresponding length value
passed to the syscall should not be touched.
This change moves the code that sets the source address (and other
fields in struct msghdr) to the nl_read_bh closure, so this code is
only executed when messages have been received on the socket.

Closes #1824.