Klibs: add firewall klib
This changeset adds a new klib that implements a network firewall.
The firewall can drop IP packets received from any network
interface, based on a set of rules defined in the manifest. Each
rule is specified as a tuple located in the `rules` array of the
`firewall` tuple in the manifest. Valid attributes for a firewall
rule are the following:
- `ip`: matches IPv4 packets, and is a tuple that can have the
following attributes:
  - `src`: matches packets based on the source IP address, which
is specified with the standard dotted notation aaa.bbb.ccc.ddd, can
be prefixed by an optional `!` character which makes the rule
match packets with an address different from the provided value,
and can be suffixed with an optional netmask (with format `/<n>`,
where `<n>` can have a value from 1 to 32) which matches packets
based on the first part of the provided address
- `ip6`: matches IPv6 packets, and is a tuple that can have the
following attributes:
  - `src`: matches packets based on the source IPv6 address, which
is specified with the standard notation for IPv6 addresses, and
similarly to its IPv4 counterpart can be prefixed with a `!`
character and suffixed with a netmask (with allowed values from 1
to 128)
- `tcp`: matches TCP packets, and is a tuple that can have the
following attributes:
  - `dest`: matches packets based on the TCP destination port
(whose value can be prefixed with a `!` character to negate the
logical comparison with the port value in a packet)
- `udp`: matches UDP packets, and is a tuple that can have a `dest`
attribute with the same format and semantics of its TCP counterpart
- `action`: indicates which action should be performed by the
firewall when a matching packet is received: allowed values are
"accept" and "drop"; if this attribute is not present, the default
action for the rule is to drop matching packets

Firewall rules are evaluated in the order they are defined, until a
matching rule (i.e. a rule where all the attributes match the
packet contents) is found and the corresponding action is executed;
if a packet does not match any rule, it is accepted.

Implements #1798.