ata_set_lba(): allow up to 256-sector I/O when LBA48 is unavailable
When not using LBA48 addressing on an ATA disk, a zero value
written to the sector count register means 256 sectors (in the same
way a zero value means 64K sectors when using LBA48 addressing).
This commit changes the sector count limit from 255 to 256 when
LBA48 is not used. In addition, the code that explicitly sets the
sector count to zero when it is 64K in LBA48 mode is being removed,
since not necessary (the two least significant bytes of the
`nsectors` variable are zero when the sector count is 64K).