ATA: honor maximum number of sectors in each I/O request
This commit adds a new function to the ATA disk driver to retrieve
the maxiumum number of sectors that can be submitted in a single
I/O request, which depends on whether LBA48 addressing is
supported. This new function is called by both the x86_64
bootloader and the kernel ATA PCI driver, which check the maximum
sector count value to split each I/O request so that it doesn't
exceed the limit.
This fixes booting on vsphere VMs with IDE disk emulation.