Makefile: move arch-specific compiler flags to $KERNCFLAGS
All compiler flags to be used for kernel and klib code, including
arch-specific flags, should be in the KERNCFLAGS Makefile variable,
so they are picked up when building both the main kernel binary and
the klibs.
This change fixes a bug caused by the compiler using the X18 CPU
register as a temporary register when compiling a klib for Aarch64.