Filesystem: change lock type from spinlock to mutex
This is being done to allow a running context to be suspended while
holding a filesystem lock (which will be needed when adding support
for remote filesystems) without risking a deadlock.
The functions to mount and unmount a filesystem are being changed
so that only the parent filesystem is locked (there is no need to
lock also the child filesystem).
The mutex_init() function is being added to allow initialization of
a struct mutex embedded in another struct (such as struct
filesystem).