pagecache_write_sg: apply fs_reserve closure before locking node
A pagecache node doesn't need to be locked when reserving
filesystem space during a file write. If an fs_reserve closure is
invoked with the node locked and suspends its context (which can
happen when support for remote filesystems is added), another
context trying to lock the node could cause a deadlock.
This change amends pagecache_write_sg() so that a node is locked
only after invoking fs_reserve().